requirements.txt
gevent==20.12.0
jsonschema==3.2.0
kafka-python==2.0.2
psycopg2-binary==2.8.6 # NOTE: https://www.psycopg.org/docs/install.html#psycopg-vs-psycopg-binary
requests==2.25.1
